% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Author_Original_Functions.R
\name{cv.concordance}
\alias{cv.concordance}
\title{Author's Implementation of Aggregated Rank Correlation to Conduct Cross-Validation for Second-Step Estimator}
\usage{
cv.concordance(lp, Xk, Dk, Ct)
}
\arguments{
\item{lp}{linear predictor for each patient (X transpose beta)}

\item{Xk}{surrogate event times}

\item{Dk}{surrogate event indicators}

\item{Ct}{censoring times for surrogate event (requires to compute inverse probability of censoring weights)}
}
\value{
Aggregated rank correlation
}
\description{
Author's Implementation of Aggregated Rank Correlation to Conduct Cross-Validation for Second-Step Estimator
}
