% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Author_Utility_Functions.R
\name{dexpexp}
\alias{dexpexp}
\title{Derivative of Expexp Function}
\usage{
dexpexp(x)
}
\arguments{
\item{x}{value}
}
\value{
d/dx exp(-exp(-x))
}
\description{
Derivative of Expexp Function
}
