% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Author_Utility_Functions.R
\name{dexpit}
\alias{dexpit}
\title{Derivative of Exponential Link Function}
\usage{
dexpit(x)
}
\arguments{
\item{x}{value}
}
\value{
dexpit(x)
}
\description{
Derivative of Exponential Link Function
}
