% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Author_Utility_Functions.R
\name{expit}
\alias{expit}
\title{Exponential Link Function (Inverse Logit)}
\usage{
expit(x)
}
\arguments{
\item{x}{value}
}
\value{
expit(x)
}
\description{
Exponential Link Function (Inverse Logit)
}
