% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Full_Implementations.R
\name{Full_Estimator_New}
\alias{Full_Estimator_New}
\title{New Implementation of Full Method for Two Surrogate Outcomes}
\usage{
Full_Estimator_New(
  data,
  dataS,
  link = expit,
  dlink = dexpit,
  inv.link = logit,
  Nperturb = 1000,
  min.lam = 1e-04,
  maxsoft = 0.5,
  Nfold = 5,
  knots = min(9, floor(m/500)),
  m = nrow(data),
  n = nrow(dataS),
  p = ncol(data$Z),
  Nlam = 100
)
}
\arguments{
\item{data}{labelled data with columns delta (event indicator), C (censoring indicator), Z (covariates)}

\item{dataS}{unlabelled data containing event time and censoring indicator for surrogate outcomes}

\item{link}{link function for score equation to be solved}

\item{dlink}{derivative of link function}

\item{inv.link}{inverse of link function}

\item{Nperturb}{number of perturbations in perturbation analysis}

\item{min.lam}{minimum value of lambda for adaptive ridge regression}

\item{maxsoft}{maximum value of lambda for adaptive ridge regression}

\item{Nfold}{number of fold in cross-validation analysis}

\item{knots}{number of knots for spline in first-step estimator}

\item{m}{size of labelled data subset}

\item{n}{size of full data}

\item{p}{number of covariates}

\item{Nlam}{number of lambdas to test in adaptive ridge regression}
}
\value{
A list containing an estimate betaSSL, and set of perturbed estimates betaSSLk for inference
}
\description{
New Implementation of Full Method for Two Surrogate Outcomes
}
