% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project_New_Functions.R
\name{init.beta.new}
\alias{init.beta.new}
\title{New Implementation of First Step Estimator based on Penalized Splines}
\usage{
init.beta.new(
  delta,
  C,
  Z,
  df = 3,
  knots = min(9, floor(length(delta)/500)),
  lambda = NULL
)
}
\arguments{
\item{delta}{event indicator}

\item{C}{censoring times}

\item{Z}{covariates (design matrix)}

\item{df}{degrees of freedom for spline}

\item{knots}{number of knots for spline}

\item{lambda}{penalty terms lambda to consider in cross-validation}
}
\value{
A list of coefficients beta corresponding to each covariate column in Z
}
\description{
New Implementation of First Step Estimator based on Penalized Splines
}
