% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Author_Original_Functions.R
\name{noise.exp.L}
\alias{noise.exp.L}
\title{Author's Original Function for applying high exponential noise to event times}
\usage{
noise.exp.L(t)
}
\arguments{
\item{t}{event times}
}
\value{
matrix with two mismeasured surrogate times
}
\description{
Author's Original Function for applying high exponential noise to event times
}
