% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Author_Original_Functions.R
\name{noise.loglinear.S}
\alias{noise.loglinear.S}
\title{Author's Original Function for applying low log-linear noise to event times}
\usage{
noise.loglinear.S(t)
}
\arguments{
\item{t}{event times}
}
\value{
matrix with two mismeasured surrogate times
}
\description{
Author's Original Function for applying low log-linear noise to event times
}
