\name{RISKmis-package}
\alias{RISKmis-package}
\alias{RISKmis}
\docType{package}
\title{
\packageTitle{RISKmis}
}
\description{
\packageDescription{RISKmis}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{RISKmis}
\packageIndices{RISKmis}
}
\author{
\packageAuthor{RISKmis}

Maintainer: \packageMaintainer{RISKmis}
}
\references{
Hou, J., Chan, S.F., Wang, X., Cai, T. Risk prediction with imperfect survival outcome information
from electronic health records. Biometrics 2021; 79, 190–202.}

