% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Author_Original_Functions.R
\name{sim.gen.rev}
\alias{sim.gen.rev}
\title{Author's Original Function for Generating Simulated Data}
\usage{
sim.gen.rev(
  n,
  beta0,
  sigmaZ,
  m,
  misT,
  Nperturb = 0,
  inv.link = logit,
  inv.h = function(x) exp(x/3) * 4
)
}
\arguments{
\item{n}{total number of observations}

\item{beta0}{true effect of each covariate}

\item{sigmaZ}{variance-covariance matrix for covariates}

\item{m}{number of labelled observations}

\item{misT}{noise function relating surrogate outcome times to true outcome times}

\item{Nperturb}{additional perturbations fo be generated (default none)}

\item{inv.link}{inverse of link function relating censoring time to probability of event}

\item{inv.h}{inverse of link function relating linear predictor to survival time}
}
\value{
A list containing two datasets: data (covariates and true outcomes) and dataS (surrogate outcomes)
}
\description{
Author's Original Function for Generating Simulated Data
}
