% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Author_Original_Functions.R
\name{Sk_sym}
\alias{Sk_sym}
\title{Author's Original Implementation of Score Function for Second Step Estimator
Corresponds to a single surrogate outcome (run one time per surrogate)}
\usage{
Sk_sym(lp, Z, Xk, Dk, Ct, K, h)
}
\arguments{
\item{lp}{linear predictor for each patient (X transpose beta)}

\item{Z}{covariates (design matrix)}

\item{Xk}{surrogate event times}

\item{Dk}{surrogate event indicators}

\item{Ct}{censoring times for surrogate event (requires to compute inverse probability of censoring weights)}

\item{K}{kernel function (e.g. dnorm)}

\item{h}{kernel bandwith}
}
\value{
Score function value for each first step estimator beta (later to be combined to correct initial estimator)
}
\description{
Author's Original Implementation of Score Function for Second Step Estimator
Corresponds to a single surrogate outcome (run one time per surrogate)
}
