% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project_New_Functions.R
\name{Sk_sym_new}
\alias{Sk_sym_new}
\title{New Implementation of Score Function for Second Step Estimator
Corresponds to a single surrogate outcome (run one time per surrogate)}
\usage{
Sk_sym_new(lp, Z, Xk, Dk, Ct, h)
}
\arguments{
\item{lp}{linear predictor for each patient (X transpose beta)}

\item{Z}{covariates (design matrix)}

\item{Xk}{surrogate event times}

\item{Dk}{surrogate event indicators}

\item{Ct}{censoring times for surrogate event (requires to compute inverse probability of censoring weights)}

\item{h}{kernel bandwith}
}
\value{
Score function value for each first step estimator beta
}
\description{
New Implementation of Score Function for Second Step Estimator
Corresponds to a single surrogate outcome (run one time per surrogate)
}
