% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Author_Original_Functions.R
\name{W_hat_adaPCA_ridge}
\alias{W_hat_adaPCA_ridge}
\title{Author's Implementation of Adaptive Ridge Regression for Second-Step Estimator}
\usage{
W_hat_adaPCA_ridge(betak, Skb, a = 2)
}
\arguments{
\item{betak}{set of first step estimators obtained from perturbation analysis}

\item{Skb}{set of score functions for first step estimators obtained from perturbation}

\item{a}{number of surrogate outcomes}
}
\value{
A list containing W.hat (optimal loading matrix), optimal lambda, penalty factor
}
\description{
Author's Implementation of Adaptive Ridge Regression for Second-Step Estimator
}
