% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Author_Utility_Functions.R
\name{wtd.mean}
\alias{wtd.mean}
\title{Weighted Mean Function}
\usage{
wtd.mean(x, wgt)
}
\arguments{
\item{x}{value}

\item{wgt}{weight}
}
\value{
weighted mean
}
\description{
Weighted Mean Function
}
