% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Author_Utility_Functions.R
\name{wtd.sd}
\alias{wtd.sd}
\title{Weighted Standard Deviation Function}
\usage{
wtd.sd(x, wgt)
}
\arguments{
\item{x}{value}

\item{wgt}{weight}
}
\value{
weighted sd
}
\description{
Weighted Standard Deviation Function
}
